Use ivy BasicURLHandler instead of okhttp variant
Why:
Okhttp was removed from SBT in 1.7 for security reasons. They switched
to using the ivy BasicURLHandler instead. This updates the plugin to
extend that class (sadly having to duplicate much of the innards).

This is required to build projects that require SBT >1.7, for example
0.0.1 fails when trying to use it to build scala 2.13.11 or 2.13.12
due to the missing class and downgrading SBT is not feasible as the
project has complex SBT integration.